avoid cyclic import as far as possible
